Mark Wolfenstein as a completed / abandoned project
As this will be my last commit to this repo, let this be the true README.

Long ago, before the Plague, in February of 2020, I was tasked with taking the [70-480](https://docs.microsoft.com/en-us/learn/certifications/exams/70-480) Microsoft exam to renew our company's Partner status or something like that. I was given a week of free time to study for the exam, and I spent it writing this little toy clone of the 1992 game, [Wolfenstein 3D](https://en.wikipedia.org/wiki/Wolfenstein_3D). I was more than prepared for the exam but this gave me an opportunity to do something creative with my time that still was technically practicing JS.

I learned to code by writing little toy programs like this, my first was Pac-Man, back in highschool, and I enjoyed this little flashback to my youth. I never licensed the code for Wolfenstein, and nearly wiped it from my hard drive at one point, but since PacMan and the others have been lost to time, I hereby preserve this horribly optimized, featureless, and fisheyed toy to the halls of history.